+++ src/ccl_halofit.c
@@ -58,7 +58,7 @@ static ccl_cosmology *create_w0eff_cosmo(double w0eff, ccl_cosmology *cosmo, int
 
   params_w0eff = ccl_parameters_create(
     cosmo->params.Omega_c, cosmo->params.Omega_b, cosmo->params.Omega_k,
-    cosmo->params.Neff, mnu, ccl_mnu_list,
+    cosmo->params.Neff, mnu, cosmo->params.N_nu_mass,
     w0eff, 0, cosmo->params.h, norm_pk,
     cosmo->params.n_s, cosmo->params.bcm_log10Mc, cosmo->params.bcm_etab,
     cosmo->params.bcm_ks, cosmo->params.mu_0, cosmo->params.sigma_0, cosmo->params.nz_mgrowth,